rename component and models
